+++ rust/crates/imperialism-core/src/ai/mod.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
 pub struct AiTradeState {
     /// Temporary stock synthesized for resources food through fuel during AI bidding.
-    pub temporary_processed_stock: ProcessedTradeCommodityTable<i16>,
+    pub temporary_processed_stock: ProcessedTradeCommodityTable<i32>,
 }
 
 /// Persistent inputs and pending work owned by retail's city interior minister.
@@ -14,15 +14,15 @@ pub struct AiTradeState {
 pub struct InteriorCivilianState {
     pub(crate) pending_recruitment: Option<CivilianUnitKind>,
     pub(crate) railhead_target: Option<TileId>,
-    pub(crate) resource_order_metrics: ResourceTable<i16>,
+    pub(crate) resource_order_metrics: ResourceTable<i32>,
     pub(crate) city_order_demand: AiCityOrderDemand,
-    pub(crate) deferred_labor_shortfall: i16,
-    pub(crate) production_deficit_by_slot: ProductionTable<i16>,
-    pub(crate) temporarily_reserved_ship_arms: i16,
-    pub(crate) railhead_priority_by_resource: ResourceTable<i16>,
-    pub(crate) exterior_need_by_resource: ResourceTable<i16>,
-    pub(crate) historical_need_by_resource: ResourceTable<i16>,
-    pub(crate) civilian_order_demand_by_resource: ResourceTable<i16>,
+    pub(crate) deferred_labor_shortfall: i32,
+    pub(crate) production_deficit_by_slot: ProductionTable<i32>,
+    pub(crate) temporarily_reserved_ship_arms: i32,
+    pub(crate) railhead_priority_by_resource: ResourceTable<i32>,
+    pub(crate) exterior_need_by_resource: ResourceTable<i32>,
+    pub(crate) historical_need_by_resource: ResourceTable<i32>,
+    pub(crate) civilian_order_demand_by_resource: ResourceTable<i32>,
     pub(crate) average_development_order_allocation: i32,
     pub(crate) pending_development_actions: Vec<PendingDevelopmentAction>,
 }
@@ -34,24 +34,24 @@ pub struct InteriorCivilianState {
 /// accepted so the next city pass can retry the remaining request.
 #[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
 pub struct AiCityOrderDemand {
-    pub(crate) training: TrainingOrderTable<i16>,
-    pub(crate) military_recruitment: MilitaryRecruitOrderTable<i16>,
-    pub(crate) civilian_recruitment: CivilianUnitTable<i16>,
-    pub(crate) ships: ShipOrderTable<i16>,
-    pub(crate) transport_capacity: i16,
-    pub(crate) expansions: ExpansionOrderTable<i16>,
-    pub(crate) population_growth: i16,
+    pub(crate) training: TrainingOrderTable<i32>,
+    pub(crate) military_recruitment: MilitaryRecruitOrderTable<i32>,
+    pub(crate) civilian_recruitment: CivilianUnitTable<i32>,
+    pub(crate) ships: ShipOrderTable<i32>,
+    pub(crate) transport_capacity: i32,
+    pub(crate) expansions: ExpansionOrderTable<i32>,
+    pub(crate) population_growth: i32,
 }
 
 impl AiCityOrderDemand {
     pub fn from_parts(
-        training: TrainingOrderTable<i16>,
-        military_recruitment: MilitaryRecruitOrderTable<i16>,
-        civilian_recruitment: CivilianUnitTable<i16>,
-        ships: ShipOrderTable<i16>,
-        transport_capacity: i16,
-        expansions: ExpansionOrderTable<i16>,
-        population_growth: i16,
+        training: TrainingOrderTable<i32>,
+        military_recruitment: MilitaryRecruitOrderTable<i32>,
+        civilian_recruitment: CivilianUnitTable<i32>,
+        ships: ShipOrderTable<i32>,
+        transport_capacity: i32,
+        expansions: ExpansionOrderTable<i32>,
+        population_growth: i32,
     ) -> Self {
         Self {
             training,
@@ -64,31 +64,31 @@ impl AiCityOrderDemand {
         }
     }
 
-    pub const fn training(&self) -> &TrainingOrderTable<i16> {
+    pub const fn training(&self) -> &TrainingOrderTable<i32> {
         &self.training
     }
 
-    pub const fn military_recruitment(&self) -> &MilitaryRecruitOrderTable<i16> {
+    pub const fn military_recruitment(&self) -> &MilitaryRecruitOrderTable<i32> {
         &self.military_recruitment
     }
 
-    pub const fn civilian_recruitment(&self) -> &CivilianUnitTable<i16> {
+    pub const fn civilian_recruitment(&self) -> &CivilianUnitTable<i32> {
         &self.civilian_recruitment
     }
 
-    pub const fn ships(&self) -> &ShipOrderTable<i16> {
+    pub const fn ships(&self) -> &ShipOrderTable<i32> {
         &self.ships
     }
 
-    pub const fn transport_capacity(&self) -> i16 {
+    pub const fn transport_capacity(&self) -> i32 {
         self.transport_capacity
     }
 
-    pub const fn expansions(&self) -> &ExpansionOrderTable<i16> {
+    pub const fn expansions(&self) -> &ExpansionOrderTable<i32> {
         &self.expansions
     }
 
-    pub const fn population_growth(&self) -> i16 {
+    pub const fn population_growth(&self) -> i32 {
         self.population_growth
     }
 }
@@ -114,15 +114,15 @@ impl InteriorCivilianState {
     pub fn from_parts(
         pending_recruitment: Option<CivilianUnitKind>,
         railhead_target: Option<TileId>,
-        resource_order_metrics: ResourceTable<i16>,
+        resource_order_metrics: ResourceTable<i32>,
         city_order_demand: AiCityOrderDemand,
-        deferred_labor_shortfall: i16,
-        production_deficit_by_slot: ProductionTable<i16>,
-        temporarily_reserved_ship_arms: i16,
-        railhead_priority_by_resource: ResourceTable<i16>,
-        exterior_need_by_resource: ResourceTable<i16>,
-        historical_need_by_resource: ResourceTable<i16>,
-        civilian_order_demand_by_resource: ResourceTable<i16>,
+        deferred_labor_shortfall: i32,
+        production_deficit_by_slot: ProductionTable<i32>,
+        temporarily_reserved_ship_arms: i32,
+        railhead_priority_by_resource: ResourceTable<i32>,
+        exterior_need_by_resource: ResourceTable<i32>,
+        historical_need_by_resource: ResourceTable<i32>,
+        civilian_order_demand_by_resource: ResourceTable<i32>,
         average_development_order_allocation: i32,
         pending_development_actions: Vec<PendingDevelopmentAction>,
     ) -> Self {
@@ -151,7 +151,7 @@ impl InteriorCivilianState {
         self.railhead_target
     }
 
-    pub const fn resource_order_metrics(&self) -> &ResourceTable<i16> {
+    pub const fn resource_order_metrics(&self) -> &ResourceTable<i32> {
         &self.resource_order_metrics
     }
 
@@ -159,31 +159,31 @@ impl InteriorCivilianState {
         &self.city_order_demand
     }
 
-    pub const fn deferred_labor_shortfall(&self) -> i16 {
+    pub const fn deferred_labor_shortfall(&self) -> i32 {
         self.deferred_labor_shortfall
     }
 
-    pub const fn production_deficit_by_slot(&self) -> &ProductionTable<i16> {
+    pub const fn production_deficit_by_slot(&self) -> &ProductionTable<i32> {
         &self.production_deficit_by_slot
     }
 
-    pub const fn temporarily_reserved_ship_arms(&self) -> i16 {
+    pub const fn temporarily_reserved_ship_arms(&self) -> i32 {
         self.temporarily_reserved_ship_arms
     }
 
-    pub const fn railhead_priority_by_resource(&self) -> &ResourceTable<i16> {
+    pub const fn railhead_priority_by_resource(&self) -> &ResourceTable<i32> {
         &self.railhead_priority_by_resource
     }
 
-    pub const fn exterior_need_by_resource(&self) -> &ResourceTable<i16> {
+    pub const fn exterior_need_by_resource(&self) -> &ResourceTable<i32> {
         &self.exterior_need_by_resource
     }
 
-    pub const fn historical_need_by_resource(&self) -> &ResourceTable<i16> {
+    pub const fn historical_need_by_resource(&self) -> &ResourceTable<i32> {
         &self.historical_need_by_resource
     }
 
-    pub const fn civilian_order_demand_by_resource(&self) -> &ResourceTable<i16> {
+    pub const fn civilian_order_demand_by_resource(&self) -> &ResourceTable<i32> {
         &self.civilian_order_demand_by_resource
     }
 
@@ -229,7 +229,7 @@ pub enum ForeignMinisterPersonality {
 }
 
 impl ForeignMinisterPersonality {
-    pub(crate) const fn initial_skill_index(self) -> i16 {
+    pub(crate) const fn initial_skill_index(self) -> i32 {
         match self {
             Self::Base | Self::Arms => 0,
             Self::Trader => 1,
